+++ Dyalect/Compiler/CompilerError.cs
@@ -136,5 +136,7 @@ public enum CompilerError
 
     FinalOnlyMethods = 266,
 
-    MethodNested = 267
+    MethodNested = 267,
+
+    AbstractOnlyMethods = 268
 }