+++ Dyalect/Parser/Scanner.cs
@@ -11,8 +11,8 @@ namespace Dyalect.Parser
 {
     partial class Scanner
     {
-	const int maxT = 104;
-	const int noSym = 104;
+	const int maxT = 105;
+	const int noSym = 105;
 
 
         static Scanner()
@@ -175,21 +175,22 @@ namespace Dyalect.Parser
 			case "when": t.kind = 79; break;
 			case "true": t.kind = 80; break;
 			case "false": t.kind = 81; break;
-			case "and": t.kind = 82; break;
-			case "get": t.kind = 83; break;
-			case "set": t.kind = 84; break;
-			case "as": t.kind = 85; break;
-			case "match": t.kind = 86; break;
-			case "not": t.kind = 88; break;
-			case "nil": t.kind = 89; break;
-			case "else": t.kind = 90; break;
-			case "many": t.kind = 91; break;
-			case "throw": t.kind = 92; break;
-			case "try": t.kind = 94; break;
-			case "catch": t.kind = 95; break;
-			case "is": t.kind = 99; break;
-			case "yields": t.kind = 102; break;
-			case "base": t.kind = 103; break;
+			case "abstract": t.kind = 82; break;
+			case "and": t.kind = 83; break;
+			case "get": t.kind = 84; break;
+			case "set": t.kind = 85; break;
+			case "as": t.kind = 86; break;
+			case "match": t.kind = 87; break;
+			case "not": t.kind = 89; break;
+			case "nil": t.kind = 90; break;
+			case "else": t.kind = 91; break;
+			case "many": t.kind = 92; break;
+			case "throw": t.kind = 93; break;
+			case "try": t.kind = 95; break;
+			case "catch": t.kind = 96; break;
+			case "is": t.kind = 100; break;
+			case "yields": t.kind = 103; break;
+			case "base": t.kind = 104; break;
 			default: break;
 		}
         }
@@ -466,9 +467,9 @@ namespace Dyalect.Parser
 			case 70:
 				{t.kind = 78; break;}
 			case 71:
-				{t.kind = 96; break;}
+				{t.kind = 97; break;}
 			case 72:
-				{t.kind = 101; break;}
+				{t.kind = 102; break;}
 			case 73:
 				recEnd = pos; recKind = 27;
 				if (ch >= '0' && ch <= '9' || ch == '_') {AddCh(); goto case 9;}
@@ -486,9 +487,9 @@ namespace Dyalect.Parser
 				else if (ch == '=') {AddCh(); goto case 66;}
 				else {t.kind = 31; break;}
 			case 76:
-				recEnd = pos; recKind = 93;
+				recEnd = pos; recKind = 94;
 				if (ch == '?') {AddCh(); goto case 62;}
-				else {t.kind = 93; break;}
+				else {t.kind = 94; break;}
 			case 77:
 				recEnd = pos; recKind = 54;
 				if (ch == '=') {AddCh(); goto case 42;}
@@ -509,9 +510,9 @@ namespace Dyalect.Parser
 				if (ch == '|') {AddCh(); goto case 88;}
 				else {t.kind = 77; break;}
 			case 82:
-				recEnd = pos; recKind = 100;
+				recEnd = pos; recKind = 101;
 				if (ch == '^') {AddCh(); goto case 89;}
-				else {t.kind = 100; break;}
+				else {t.kind = 101; break;}
 			case 83:
 				recEnd = pos; recKind = 61;
 				if (ch == '>') {AddCh(); goto case 90;}
@@ -522,22 +523,22 @@ namespace Dyalect.Parser
 				if (ch == '=') {AddCh(); goto case 67;}
 				else {t.kind = 51; break;}
 			case 85:
-				recEnd = pos; recKind = 87;
+				recEnd = pos; recKind = 88;
 				if (ch == '.') {AddCh(); goto case 70;}
 				else if (ch == '<') {AddCh(); goto case 72;}
-				else {t.kind = 87; break;}
+				else {t.kind = 88; break;}
 			case 86:
 				recEnd = pos; recKind = 68;
 				if (ch == '<') {AddCh(); goto case 91;}
 				else {t.kind = 68; break;}
 			case 87:
-				recEnd = pos; recKind = 98;
+				recEnd = pos; recKind = 99;
 				if (ch == '&') {AddCh(); goto case 92;}
-				else {t.kind = 98; break;}
+				else {t.kind = 99; break;}
 			case 88:
-				recEnd = pos; recKind = 97;
+				recEnd = pos; recKind = 98;
 				if (ch == '|') {AddCh(); goto case 93;}
-				else {t.kind = 97; break;}
+				else {t.kind = 98; break;}
 			case 89:
 				if (ch == '^') {AddCh(); goto case 94;}
 				else {goto case 0;}