+++ Docs/Changes.md
@@ -4,6 +4,7 @@
   * Change: _General code refactoring._
   * Add: _Add built-in `instanceMember` and `staticMember` functions_ ([Issue #937](https://github.com/vorov2/dyalect/issues/937), [docs](https://github.com/vorov2/dyalect/wiki/Standard-functions#instanceMember)).
   * Add: _Add support for underscores in numeric literals_ ([Issue #942](https://github.com/vorov2/dyalect/issues/942)).
+  * Add: _Consider abstract methods and properties_ ([Issue #945](https://github.com/vorov2/dyalect/issues/945), [docs](https://github.com/vorov2/dyalect/wiki/Functions#abstract)).
   * Fix: _Dya is unable to find config file_ ([Issue #943](https://github.com/vorov2/dyalect/issues/943)).
   * Fix: _Lazy variables don't work correctly in interactive_ ([Issue #944](https://github.com/vorov2/dyalect/issues/944)).